+++ frontend/src/components/Login/Login.css
@@ -1,19 +1,15 @@
 .panel{
-    width:40%;
+    width:50%;
     margin: auto;
     padding: auto;
     padding-top: 3%;
     text-align: center;
   }
   #logo{
-    width: 20%;
+    width: 5%;
+    padding-bottom: 5%;
   }
-  img{
-    width: 1%;    
-    margin-bottom: 5%;
-    margin-left: auto;
-    margin-right: auto;
-  }
+ 
   .forgot{
     width: 100%; 
     height: 30px;
